Changed logging to use uints. File re-organization [See details]
Logging now uses uint32 for log level. 
Logging options added to config file.
Re-organized files and code because otherwise it caused cyclic dependency when using the Config type for the `Convert` (now `copyTable`) function.
DB opening moved out from `Convert` function.
CopyToX functions no longer in the same files as Config.
